Removed `int` cast to consistently use `long`. 🩺

+++ clients/da-vinci-client/src/main/java/com/linkedin/davinci/config/VeniceServerConfig.java
@@ -545,7 +545,7 @@ public class VeniceServerConfig extends VeniceClusterConfig {
   private final boolean aaWCWorkloadParallelProcessingEnabled;
   private final int aaWCWorkloadParallelProcessingThreadPoolSize;
   private final boolean isGlobalRtDivEnabled;
-  private final int pubSubConsumerWaitAfterUnsubscribeTimeoutMs;
+  private final long pubSubConsumerWaitAfterUnsubscribeTimeoutMs;
 
   public VeniceServerConfig(VeniceProperties serverProperties) throws ConfigurationException {
     this(serverProperties, Collections.emptyMap());
@@ -914,7 +914,7 @@ public class VeniceServerConfig extends VeniceClusterConfig {
         serverProperties.getBoolean(SERVER_AA_WC_WORKLOAD_PARALLEL_PROCESSING_ENABLED, false);
     aaWCWorkloadParallelProcessingThreadPoolSize =
         serverProperties.getInt(SERVER_AA_WC_WORKLOAD_PARALLEL_PROCESSING_THREAD_POOL_SIZE, 8);
-    pubSubConsumerWaitAfterUnsubscribeTimeoutMs = serverProperties.getInt(
+    pubSubConsumerWaitAfterUnsubscribeTimeoutMs = serverProperties.getLong(
         SERVER_WAIT_AFTER_UNSUBSCRIBE_TIMEOUT_MS,
         KafkaConsumerService.DEFAULT_WAIT_AFTER_UNSUBSCRIBE_TIMEOUT_MS);
   }
@@ -1155,7 +1155,7 @@ public class VeniceServerConfig extends VeniceClusterConfig {
     return pubSubConsumerPollRetryBackoffMs;
   }
 
-  public int getPubSubConsumerWaitAfterUnsubscribeTimeoutMs() {
+  public long getPubSubConsumerWaitAfterUnsubscribeTimeoutMs() {
     return pubSubConsumerWaitAfterUnsubscribeTimeoutMs;
   }
