+++ clients/da-vinci-client/src/main/java/com/linkedin/davinci/kafka/consumer/TopicWiseKafkaConsumerService.java
@@ -44,7 +44,7 @@ public class TopicWiseKafkaConsumerService extends KafkaConsumerService {
       final MetricsRepository metricsRepository,
       final String kafkaClusterAlias,
       final long sharedConsumerNonExistingTopicCleanupDelayMS,
-      final int consumerWaitAfterUnsubscribeTimeoutMs,
+      final long consumerWaitAfterUnsubscribeTimeoutMs,
       final TopicExistenceChecker topicExistenceChecker,
       final boolean liveConfigBasedKafkaThrottlingEnabled,
       final PubSubMessageDeserializer pubSubDeserializer,
